+++ lib/src/components/scorecard/src/scorecard_bar.dart
@@ -52,8 +52,9 @@ class ScorecardBarDirective implements OnInit, OnDestroy, AfterViewChecked {
   @override
   void ngOnInit() {
     _disposer.addDisposable(_domService.scheduleRead(_readElement));
-    _disposer.addDisposable(
-        _domService.trackLayoutChange(() => currentClientSize, (_) {
+    _disposer.addDisposable(_domService.trackLayoutChange(
+        () => currentClientSize.toString() + ' ' + currentScrollSize.toString(),
+        (_) {
       _readElement(windowResize: true);
       _refreshController.add(true);
     }, runInAngularZone: true));